Build navigation channel buttons in Channels List
Three main types of channel lists separated in sections:

- Starred
- Regular
- Direct Messages

For proper a11y, each group is a navigation landmark.

This is the first usage of `<h3>` heading, so now is a good time to
talk about heading architectural decisions  for screen reader
discoverability and navigation:

- `<h1>`: team name
- `<h2>`: current channel name that will be in Channel Header Component
(not yet implemented)
- `<h3>`: channels list section types (e.g. starred, channels, direct
messages)
- `<h4>`: date container containing messages in Message Pane Component
(not yet implemented, still pondering)

This heading architecture is subject to change, but this is the initial
thinking.