Support for basic Component-Based Focus Management
By using two familiar yet different keys (`Enter` and `Escape`), users
will be able to press `Enter` to begin interacting with a component and
`Escape` to stop interacting with a component.

The main motivation here is to allow `tab` to be more effective in a
rich app where the user can jump from one section (component) to
another quickly instead of tabbing through each and every interactive
element in the tab flow.