Handle deeper levels of Component-Based Focus Management
Previously, this only allowed the main components (i.e. Channels List,
Channel Header) to get focus and when interacting with it, all
interactive child elements become focusable.

Now, this supports deeper levels of focusable components.

This introduces a new term:

Spotlight Component: the current component the user is interacting
with. When the user presses `Enter` on a focusable component, that
component becomes the Spotlight Component. There can only be one
Spotlight Component active at a time.

This updates the Channels List Component to support deeper levels of
focusable components.