Make `Escape` put focus on Spotlight Component and crown the new parent Spotlight Component
We need to keep track of Spotlight Components in a state
`this.state.focus` as we interact with them such that we have a living
stack to traverse when we stop interacting with them.

Convert ConversationsList and Navigation Component from pure components
to to logical components because we need to access `this.refs` - our
`this.state.focus` stack is an array of Elements.